backend/requirements.txt
flask
flask-cors
flask-bcrypt
flask-jwt-extended
mysql-connector-python
python-dotenv
